refactor: CommandCondition command 패키지로 이동
