+++ src/main/java/chess/domain/command/CommandCondition.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.command;
 
 import java.util.List;
