+++ src/main/java/chess/domain/command/CommandExecutor.java
@@ -1,8 +1,6 @@
 package chess.domain.command;
 
 
-import chess.domain.CommandCondition;
-
 @FunctionalInterface
 public interface CommandExecutor {
     void execute(CommandCondition commandCondition);
